# MM-50435: Improve data retention queries on MySQL