@agnivade 2023-02-10
@streamer45 - Welcome back. I could use a second set of eyes here.

Context: this is a from a customer escalation. I added you to the channel: https://community.mattermost.com/private-core/pl/y6jrawezttn8xxecbewf86so8a