@agnivade 2023-02-10
For all other `DeleteOrphanedRows` query, there is no material difference in the query plan. MySQL correctly does a nested antijoin if the selecting column and the joining column is the same.