@agnivade 2023-02-10
> Code looks good :) But there's something I don't understand: in Postgres, the cost for the NOT-IN query is 1000, while the cost of the LEFT-JOIN one is ~6000. 

1000 is just the startup cost, not the full cost. See `1000.00..9419975.49`. The total estimated cost is `9419975.49` and the total cost for the left join is `655145.99`. So an order of magnitude lower.

Hope that clarifies. See [this](https://www.postgresql.org/docs/current/using-explain.html#USING-EXPLAIN-BASICS) for start-up cost and total cost.