@agnivade 2023-02-10
It needs to filter on the `posts.ChannelId` right? That seems like the right index to me. :thinking: There is no other dedicated index for channelId.