@streamer45 2023-02-10
> It needs to filter on the `posts.ChannelId` right? That seems like the right index to me. thinking There is no other dedicated index for channelId.

I was referring to the index presumably used on `Channels` table. The one on `Posts` is fine but what is `Channels.UpdateAt` going to do exactly?