@agnivade 2023-02-10
Yep! I could see some improvement in the first query, but not that much in the second. But it makes sense to use PRIMARY here. Nice catch.