@agnivade 2023-02-10
Ok, now I am confused. When I am deleting some channel rows to actually get some output from the queries, I can see the old queries perform faster now. _Although_ the cost in new queries is still lower (by an order of magnitude).

Given this is MySQL, I am not entirely sure if this is going to bomb or succeed.

Any thoughts @streamer45 @agarciamontoro ?