@agarciamontoro 2023-02-10
So if there are no orphaned rows, the NOT-IN query is faster, but if there are orphaned rows, the LEFT-JOIN query is faster?