@agnivade 2023-02-10
That seems to be what I am seeing. But the query plan is faster in both cases for new query. Let me include the DELETE and check as well.