@agnivade 2023-02-10
With DELETE, the costs are same orders of magnitude, but still the `LEFT JOIN` has higher cost. I am logging off for the day, but let me know if you guys have any thoughts.