@streamer45 2023-02-11
From what I could quickly check it's working as expected on my end (Aurora MySQL `5.7.12`), with the new query performing roughly twice as fast. 

I tested on purely synthetic data (100K channels, 1M posts) but given the query is taken in isolation I am thinking it shouldn't matter that much.

### Old query

```sql
DELETE
FROM Posts
WHERE Id IN
    (SELECT *
     FROM
       (SELECT Posts.Id
        FROM Posts
        LEFT JOIN Channels ON Posts.ChannelId = Channels.Id
        WHERE Channels.Id IS NULL
        LIMIT 3000) AS A);
 ```

```
+----+--------------------+------------+------------+----------------+---------------+--------------------------------+---------+----------------------+--------+----------+--------------------------------------+
| id | select_type        | table      | partitions | type           | possible_keys | key                            | key_len | ref                  | rows   | filtered | Extra                                |
+----+--------------------+------------+------------+----------------+---------------+--------------------------------+---------+----------------------+--------+----------+--------------------------------------+
|  1 | DELETE             | Posts      | NULL       | ALL            | NULL          | NULL                           | NULL    | NULL                 | 953336 |   100.00 | Using where                          |
|  2 | DEPENDENT SUBQUERY | <derived3> | NULL       | index_subquery | <auto_key0>   | <auto_key0>                    | 106     | func                 |    300 |   100.00 | Using index                          |
|  3 | DERIVED            | Posts      | NULL       | index          | NULL          | idx_posts_channel_id_update_at | 116     | NULL                 | 953336 |   100.00 | Using index                          |
|  3 | DERIVED            | Channels   | NULL       | eq_ref         | PRIMARY       | PRIMARY                        | 106     | test.Posts.ChannelId |      1 |   100.00 | Using where; Not exists; Using index |
+----+--------------------+------------+------------+----------------+---------------+--------------------------------+---------+----------------------+--------+----------+--------------------------------------+
```

#### No orphans

```
mysql> DELETE FROM Posts WHERE Id IN (SELECT * FROM (SELECT Posts.Id FROM Posts LEFT JOIN Channels ON Posts.ChannelId = Channels.Id WHERE Channels.Id IS NULL LIMIT 3000) AS A);
Query OK, 0 rows affected (7.17 sec)
```

#### With orphans (> 3000)

```
mysql> DELETE FROM Posts WHERE Id IN (SELECT * FROM (SELECT Posts.Id FROM Posts LEFT JOIN Channels ON Posts.ChannelId = Channels.Id WHERE Channels.Id IS NULL LIMIT 3000) AS A);
Query OK, 3000 rows affected (7.71 sec)
```

### New query

```sql
DELETE
FROM Posts
WHERE Id IN
    (SELECT *
     FROM
       (SELECT Posts.Id
        FROM Posts
        WHERE Posts.ChannelId NOT IN
            (SELECT Id
             FROM Channels USE INDEX(PRIMARY))
        LIMIT 3000) AS A);
 ```

```
+----+--------------------+------------+------------+----------------+---------------+--------------------------------+---------+------+--------+----------+--------------------------+
| id | select_type        | table      | partitions | type           | possible_keys | key                            | key_len | ref  | rows   | filtered | Extra                    |
+----+--------------------+------------+------------+----------------+---------------+--------------------------------+---------+------+--------+----------+--------------------------+
|  1 | DELETE             | Posts      | NULL       | ALL            | NULL          | NULL                           | NULL    | NULL | 953336 |   100.00 | Using where              |
|  2 | DEPENDENT SUBQUERY | <derived3> | NULL       | index_subquery | <auto_key0>   | <auto_key0>                    | 106     | func |    300 |   100.00 | Using index              |
|  3 | DERIVED            | Posts      | NULL       | index          | NULL          | idx_posts_channel_id_update_at | 116     | NULL | 953336 |   100.00 | Using where; Using index |
|  4 | SUBQUERY           | Channels   | NULL       | index          | PRIMARY       | PRIMARY                        | 106     | NULL |  99483 |   100.00 | Using index              |
+----+--------------------+------------+------------+----------------+---------------+--------------------------------+---------+------+--------+----------+--------------------------+
```

#### No orphans

```
mysql> DELETE FROM Posts WHERE Id IN (SELECT * FROM (SELECT Posts.Id FROM Posts WHERE Posts.ChannelId NOT IN (SELECT Id FROM Channels USE INDEX(PRIMARY)) LIMIT 3000) AS A);
Query OK, 0 rows affected (3.17 sec)
```

#### With orphans (> 3000)

```
mysql> DELETE FROM Posts WHERE Id IN (SELECT * FROM (SELECT Posts.Id FROM Posts WHERE Posts.ChannelId NOT IN (SELECT Id FROM Channels USE INDEX(PRIMARY)) LIMIT 3000) AS A);
Query OK, 3000 rows affected (4.17 sec)
```